writed the BSP function

+++ day02/ex03/Point.cpp
@@ -1,7 +1,6 @@
 #include "Point.hpp"
 
-Point::Point(void)
-{ }
+Point::Point(void) { }
 
 Point::Point(const Fixed &_x, const Fixed &_y)
 : x(_x), y(_y)
@@ -32,8 +31,8 @@ Point   Point::operator+(const Point &obj) const
     return Point(x + obj.x, y + obj.y);
 }
 
-int    Point::checkLine(const Point &p1, const Point &p2) const
+float   Point::checkLine(const Point &p1, const Point &p2) const
 {
-    return (p1.x - x) * (p2.y - p1.y) - (p2.x - p1.x) * (p1.y - y);
+    return ((p1.x - x) * (p2.y - p1.y) - (p2.x - p1.x) * (p1.y - y)).toFloat();
 }
